Porting resnet
